Integration with Economic Strategy Manager
